refactoring des membres

+++ app/Http/Controllers/Admin/MembreController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 class MembreController extends Controller
 {
@@ -19,8 +20,10 @@ class MembreController extends Controller
      */
     public function index()
     {
-        $admins = User::all();
-        return view('admin.admin.index',compact('admins'));
+       if (Auth::user()->isAdmin == 1) {
+            $admins = User::all();
+            return view('admin.admin.index',compact('admins'));
+       }
     }
 
     /**
@@ -41,16 +44,23 @@ class MembreController extends Controller
      */
      public function store(Request $request)
     {
+        if (Auth::user()->isAdmin == 1) {
             $this->validate($request,[
                 'name' => 'required|string',
                 'email' => 'required|unique:users',
                 'phone' => 'required|unique:users|numeric',
                 'password' => 'required|min:6|string',
             ]);
-            $request['password'] = Hash::make($request->password);
-            $admin = User::create($request->all());
+            $admin = new User();
+            $admin->name = $request->name;
+            $admin->email = $request->email;
+            $admin->phone = $request->phone;
+            $admin->status = $request->status;
+            $admin->isAdmin = $request->admin;
+            $admin->password = Hash::make($request->password);
             $admin->save();
-            return redirect(route('membre.index'));
+            return redirect(route('membre.index'))->with('success','Votre admin a ete ajoute');
+         }
     }
 
     /**
@@ -72,8 +82,10 @@ class MembreController extends Controller
      */
     public function edit($id)
     {
+        if (Auth::user()->isAdmin == 1) {
             $admins = User::find($id);
             return view('admin.admin.edit',compact('admins'));
+        }
     }
 
     /**
@@ -85,9 +97,11 @@ class MembreController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->status? : $request['status'] = 0 ;
-        $user = User::where('id',$id)->update($request->except('_token','_method'));
-        return redirect()->route('membre.index');
+        if (Auth::user()->isAdmin == 1) {
+            $request->status? : $request['status'] = 0 ;
+            $user = User::where('id',$id)->update($request->except('_token','_method'));
+            return redirect()->route('membre.index')->with('success','La modification a ete enregistre');
+        }
     }
 
     /**
@@ -98,7 +112,9 @@ class MembreController extends Controller
      */
     public function destroy($id)
     {
-        User::where('id',$id)->delete();
-        return back();
+        if (Auth::user()->isAdmin == 1) {
+            User::where('id',$id)->delete();
+            return back()->with('success','Ce membre a ete supprime');
+        }
     }
 }
